feat: prediction get all by user

+++ UniversityStudentTracker.API/Models/DTO/Prediction/PredictionResponseDto.cs
@@ -2,6 +2,6 @@ namespace UniversityStudentTracker.API.Models.DTO.Prediction;
 
 public class PredictionResponseDto
 {
-    public decimal PredictedGrade { get; set; }
+    public double PredictedGrade { get; set; }
     public int PredictedKnowledgeLevel { get; set; }
 }
